# perf(webidl): optimize createDictionaryConverter()