On a benchmark constructing `Response`s with headers this shaves off 25%:
```
Before:
resp_w_h:            	n = 1000000, dt = 2.284s, r = 437828/s, t = 2283ns/op
After:
resp_w_h:            	n = 1000000, dt = 1.725s, r = 579710/s, t = 1725ns/op
```