@kurtextrem 2021-10-05
Interesting, if I micro-benchmark this in V8 9.6.11 (Chrome v96), I get a majorly faster spread operation: https://esbench.com/bench/615c7bcb6c89f600a5701202